requirements.txt
#requirements.txt
pandas
scikit-learn
pyyaml
dvc

